build: move jest-circus/ facade to jest/
