+++ detox/runners/jest/adapter.js
@@ -1,1 +1,3 @@
-require('../deprecation');
+/** @deprecated */
+/** @type {never} */
+module.exports = require('../deprecation');
