+++ detox/runners/jest/index.js
@@ -1,0 +1,9 @@
+const DetoxCircusEnvironment = require('./environment');
+const SpecReporter = require('./listeners/SpecReporter');
+const WorkerAssignReporter = require('./listeners/WorkerAssignReporter');
+
+module.exports = {
+  DetoxCircusEnvironment,
+  SpecReporter,
+  WorkerAssignReporter,
+};
