+++ detox/runners/jest/reporter.js
@@ -1,0 +1,1 @@
+module.exports = require('./reporters/DetoxReporter');
