+++ detox/runners/jest/utils/hasTimedOut.js
