+++ detox/runners/jest/utils/stdout.js
